backend/requirements.txt
Django==4.2
djangorestframework
Markdown
django-filter
django-cors-headers
djangorestframework-simplejwt